+++ ceammc/ext/src/array/array_grainer.h
@@ -63,6 +63,7 @@ public:
     void m_spread(t_symbol* s, const AtomListView& lv);
     void m_shuffle(t_symbol* s, const AtomListView& lv);
     void m_defer(t_symbol* s, const AtomListView& lv);
+    void m_reverse(t_symbol* s, const AtomListView& lv);
 
     const GrainCloud& cloud() const { return cloud_; }