+++ ceammc/CHANGELOG.md
@@ -39,6 +39,7 @@
   - [spread( method added to \[array.grainer~\] - to uniformly spread grain over specified duration
   - [shuffle( method added to \[array.grainer~\] - to shuffle grains
   - [defer( method added to \[array.grainer~\] - to execute messages on grain finish
+  - [reverse( method added to \[array.grainer~\] - to reverse grains positions
 - misc:
   - conv.lin2exp: list support added
   - conv.lin2curve: list support added